pyproject.toml
[tool.poetry]
name = "financetracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
uvicorn = "^0.30.2"
SQLAlchemy = "^2.0.31"
alembic = "^1.13.2"
python-decouple = "^3.8"
asyncpg = "^0.29.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
